created new arrayadaptor for qr, renamed qrcode manager to qrcodegenerator, and made qr code activity more complete
