+++ app/src/main/java/com/example/trialio/adapters/ArrayAdapterQR.java
@@ -1,0 +1,4 @@
+package com.example.trialio.adapters;
+
+public class ArrayAdapterQR {
+}
