--- app/src/main/java/com/example/trialio/controllers/QRCodeManager.java
@@ -1,4 +1,0 @@
-package com.example.trialio.controllers;
-
-public class QRCodeManager {
-}
