--- app/src/main/java/com/example/trialio/models/QRCode.java
@@ -1,4 +1,0 @@
-package com.example.trialio.models;
-
-public class QRCode {
-}
