+++ app/src/main/java/com/example/trialio/models/QrCode.java
@@ -1,0 +1,10 @@
+package com.example.trialio.models;
+
+import androidx.appcompat.app.AppCompatActivity;
+import android.os.Bundle;
+
+public class QrCode extends AppCompatActivity {
+
+
+
+}
